fix: mypy errors
