+++ src/asammdf/blocks/mdf_v4.py
@@ -87,7 +87,6 @@ from .types import (
     ChannelsType,
     CompressionType,
     DbcFileType,
-    RasterType,
     StrPath,
 )
 from .utils import (
@@ -374,7 +373,7 @@ class MDF4(MDF_Common[Group]):
 
         self._column_storage = False
 
-        self._units_map = {}
+        self._units_map: dict[int, str] = {}
 
         super().__init__(kwargs.get("raise_on_multiple_occurrences", GLOBAL_OPTIONS["raise_on_multiple_occurrences"]))
 
@@ -881,6 +880,7 @@ class MDF4(MDF_Common[Group]):
 
             if filter_channels:
                 if mapped:
+                    stream = typing.cast(mmap.mmap, stream)
                     (
                         id_,
                         links_nr,
@@ -1908,6 +1908,7 @@ class MDF4(MDF_Common[Group]):
         uses_ld = False
 
         if mapped:
+            stream = typing.cast(mmap.mmap, stream)
             if address:
                 if address + COMMON_SHORT_SIZE > self.file_limit:
                     handle_incomplete_block(address, self.original_name)
@@ -1974,6 +1975,7 @@ class MDF4(MDF_Common[Group]):
         READ_CHUNK_SIZE = min(READ_CHUNK_SIZE, total_size)
 
         if mapped:
+            stream = typing.cast(mmap.mmap, stream)
             if original_address := address:
                 if address + COMMON_SHORT_SIZE > self.file_limit:
                     return handle_incomplete_block(original_address, self.original_name)
@@ -5010,6 +5012,7 @@ class MDF4(MDF_Common[Group]):
             # first add the signals in the simple signal list
             if sig_type == v4c.SIGNAL_TYPE_SCALAR:
                 # compute additional byte offset for large records size
+                array: NDArray[Any]
                 if sig.dtype.kind == "O":
                     array = encode(sig.array.astype(str), "utf-8")
                 else:
@@ -6855,13 +6858,236 @@ class MDF4(MDF_Common[Group]):
 
         return data, file_path, md5_sum
 
+    def _cut_inplace(
+        self,
+        start: float | None = None,
+        stop: float | None = None,
+        whence: int = 0,
+        progress: Any | None = None,
+    ) -> None:
+        """Cut `MDF4`. `start` and `stop` are absolute values or values relative
+        to the first timestamp depending on the `whence` argument. This
+        function is a lot faster than `cut`.
+
+        .. versionadded:: 8.7.0
+
+        Parameters
+        ----------
+        start : float, optional
+            Start time; default is None. If None, the start of the measurement
+            is used.
+        stop : float, optional
+            Stop time; default is None. If None, the end of the measurement is
+            used.
+        whence : int, default 0
+            How to search for the start and stop values.
+
+            * 0 : absolute
+            * 1 : relative to first timestamp
+
+        Returns
+        -------
+        out : MDF4
+            New `MDF4` object if the inplace cut cannot be done, or `self` if
+            the inplace cut is possible
+
+        """
+
+        if start is None and stop is None:
+            return
+
+        if whence == 1:
+            timestamps: list[float] = []
+            for gp_idx in self.virtual_groups:
+                master = self.get_master(gp_idx, record_offset=0, record_count=1)
+                if master.size:
+                    timestamps.append(master[0])
+
+            if timestamps:
+                first_timestamp = np.amin(timestamps)
+            else:
+                first_timestamp = 0
+
+            if start is not None:
+                start += first_timestamp
+            if stop is not None:
+                stop += first_timestamp
+
+        groups_nr = len(self.groups)
+
+        if progress is not None:
+            if callable(progress):
+                progress(0, groups_nr)
+            else:
+                progress.signals.setValue.emit(0)
+                progress.signals.setMaximum.emit(groups_nr)
+
+        for i, group in enumerate(self.groups):
+            channel_group = group.channel_group
+            record_size = channel_group.samples_byte_nr + channel_group.invalidation_bytes_nr
+
+            stream: FileLike | mmap.mmap | tempfile._TemporaryFileWrapper[bytes]
+            if group.data_location == v4c.LOCATION_ORIGINAL_FILE:
+                stream = typing.cast(FileLike | mmap.mmap, self._file)
+            else:
+                stream = self._tempfile
+
+            out_seek = self._tempfile.seek
+            out_tell = self._tempfile.tell
+            out_write = self._tempfile.write
+
+            read = stream.read
+            seek = stream.seek
+
+            new_blocks = []
+            new_cycles_nr = 0
+
+            for j, info in enumerate(group.data_blocks):
+                if progress and progress.stop:
+                    raise Terminated
+                (
+                    address,
+                    original_size,
+                    compressed_size,
+                    block_type,
+                    param,
+                    block_limit,
+                ) = (
+                    info.address,
+                    typing.cast(int, info.original_size),
+                    info.compressed_size,
+                    info.block_type,
+                    info.param,
+                    info.block_limit,
+                )
+
+                seek(address)
+                new_data = read(typing.cast(int, compressed_size))
+
+                match block_type:
+                    case v4c.DZ_BLOCK_DEFLATE:
+                        new_data = decompress(new_data, bufsize=original_size)
+                    case v4c.DZ_BLOCK_TRANSPOSED:
+                        new_data = decompress(new_data, bufsize=original_size)
+                        cols = typing.cast(int, param)
+                        lines = original_size // cols
+                        matrix_size = lines * cols
+
+                        if matrix_size != original_size:
+                            new_data = (
+                                frombuffer(new_data[:matrix_size], dtype=uint8)
+                                .reshape((cols, lines))
+                                .T.ravel()
+                                .tobytes()
+                                + new_data[matrix_size:]
+                            )
+                        else:
+                            new_data = frombuffer(new_data, dtype=uint8).reshape((cols, lines)).T.ravel().tobytes()
+
+                    case v4c.DZ_BLOCK_LZ:
+                        new_data = lz_decompress(new_data)
+
+                if block_limit is not None:
+                    new_data = new_data[:block_limit]
+
+                count = len(new_data) // record_size
+
+                fragment = Fragment(
+                    new_data,
+                    0,
+                    count,
+                    None,
+                )
+
+                master = self.get_master(i, data=fragment, one_piece=True)
+
+                if not len(master):
+                    continue
+
+                needs_cutting = True
+
+                if start is None:
+                    start_index = 0
+                    if master[0] > stop:
+                        break
+                    else:
+                        fragment_stop = min(stop, master[-1])
+                        stop_index = np.searchsorted(master, fragment_stop, side="right")
+                        if stop_index == len(master):
+                            needs_cutting = False
+
+                elif stop is None:
+                    start_index = 0
+                    if master[-1] < start:
+                        continue
+                    else:
+                        fragment_start = max(start, master[0])
+                        start_index = np.searchsorted(master, fragment_start, side="left")
+                        stop_index = len(master)
+                        if start_index == 0:
+                            needs_cutting = False
+                else:
+                    if master[0] > stop:
+                        break
+                    elif master[-1] < start:
+                        continue
+                    else:
+                        fragment_start = max(start, master[0])
+                        start_index = np.searchsorted(master, fragment_start, side="left")
+                        fragment_stop = min(stop, master[-1])
+                        stop_index = np.searchsorted(master, fragment_stop, side="right")
+                        if start_index == 0 and stop_index == len(master):
+                            needs_cutting = False
+
+                if needs_cutting:
+                    data = new_data[start_index * record_size : stop_index * record_size]
+                    count = stop_index - start_index
+
+                else:
+                    data = new_data
+
+                out_seek(0, 2)
+
+                raw_size = len(data)
+                data = lz_compress(data, store_size=True)
+
+                size = len(data)
+                out_seek(0, 2)
+                data_address = out_tell()
+                out_write(data)
+
+                info = DataBlockInfo(
+                    address=data_address,
+                    block_type=v4c.DZ_BLOCK_LZ,
+                    original_size=raw_size,
+                    compressed_size=size,
+                    param=0,
+                )
+
+                new_blocks.append(info)
+                new_cycles_nr += count
+
+            channel_group.cycles_nr = new_cycles_nr
+            group.data_blocks = new_blocks
+            group.data_location = v4c.LOCATION_TEMPORARY_FILE
+
+            if progress is not None:
+                if callable(progress):
+                    progress(i + 1, groups_nr)
+                else:
+                    progress.signals.setValue.emit(i + 1)
+
+                    if progress.stop:
+                        print("return terminated")
+                        raise Terminated
+
     @overload
     def get(
         self,
         name: str | None = ...,
         group: int | None = ...,
         index: int | None = ...,
-        raster: RasterType | None = ...,
+        raster: float | None = ...,
         samples_only: Literal[False] = ...,
         data: Fragment | None = ...,
         raw: bool = ...,
@@ -6877,7 +7103,7 @@ class MDF4(MDF_Common[Group]):
         name: str | None = ...,
         group: int | None = ...,
         index: int | None = ...,
-        raster: RasterType | None = ...,
+        raster: float | None = ...,
         *,
         samples_only: Literal[True],
         data: Fragment | None = ...,
@@ -6894,7 +7120,7 @@ class MDF4(MDF_Common[Group]):
         name: str | None = ...,
         group: int | None = ...,
         index: int | None = ...,
-        raster: RasterType | None = ...,
+        raster: float | None = ...,
         *,
         samples_only: Literal[True],
         data: Fragment | None = ...,
@@ -6911,7 +7137,7 @@ class MDF4(MDF_Common[Group]):
         name: str | None = ...,
         group: int | None = ...,
         index: int | None = ...,
-        raster: RasterType | None = ...,
+        raster: float | None = ...,
         *,
         samples_only: Literal[True],
         data: Fragment | None = ...,
@@ -6928,7 +7154,7 @@ class MDF4(MDF_Common[Group]):
         name: str | None = ...,
         group: int | None = ...,
         index: int | None = ...,
-        raster: RasterType | None = ...,
+        raster: float | None = ...,
         samples_only: bool = ...,
         data: Fragment | None = ...,
         raw: bool = ...,
@@ -6943,7 +7169,7 @@ class MDF4(MDF_Common[Group]):
         name: str | None = None,
         group: int | None = None,
         index: int | None = None,
-        raster: RasterType | None = None,
+        raster: float | None = None,
         samples_only: bool = False,
         data: Fragment | None = None,
         raw: bool = False,
@@ -7283,7 +7509,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[tuple[int, int]],
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: Literal[False],
         record_offset: int,
@@ -7300,7 +7526,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[tuple[int, int]],
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: Literal[True],
         record_offset: int,
@@ -7317,7 +7543,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[tuple[int, int]],
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: Literal[False],
         record_offset: int,
@@ -7334,7 +7560,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[tuple[int, int]],
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: Literal[True],
         record_offset: int,
@@ -7351,7 +7577,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[tuple[int, int]],
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: bool,
         record_offset: int,
@@ -7367,7 +7593,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[tuple[int, int]],
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: bool,
         record_offset: int,
@@ -7589,7 +7815,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[ChannelArrayBlock],
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: Literal[False],
         record_offset: int,
@@ -7605,7 +7831,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[ChannelArrayBlock],
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: Literal[True],
         record_offset: int,
@@ -7621,7 +7847,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[ChannelArrayBlock],
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: Literal[False],
         record_offset: int,
@@ -7637,7 +7863,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[ChannelArrayBlock],
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: Literal[True],
         record_offset: int,
@@ -7653,7 +7879,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[ChannelArrayBlock],
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: bool,
         record_offset: int,
@@ -7668,7 +7894,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[ChannelArrayBlock],
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: bool,
         record_offset: int,
@@ -8030,7 +8256,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[ChannelArrayBlock] | list[tuple[int, int]] | None,
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: Literal[False],
         record_offset: int,
@@ -8047,7 +8273,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[ChannelArrayBlock] | list[tuple[int, int]] | None,
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: Literal[True],
         record_offset: int,
@@ -8064,7 +8290,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[ChannelArrayBlock] | list[tuple[int, int]] | None,
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: Literal[False],
         record_offset: int,
@@ -8081,7 +8307,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[ChannelArrayBlock] | list[tuple[int, int]] | None,
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: Literal[True],
         record_offset: int,
@@ -8098,7 +8324,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[ChannelArrayBlock] | list[tuple[int, int]] | None,
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: bool,
         record_offset: int,
@@ -8114,7 +8340,7 @@ class MDF4(MDF_Common[Group]):
         group_index: int,
         channel_index: int,
         dependency_list: list[ChannelArrayBlock] | list[tuple[int, int]] | None,
-        raster: RasterType | None,
+        raster: float | None,
         data: Fragment | None,
         ignore_invalidation_bits: bool,
         record_offset: int,
