+++ types-requirements.txt
@@ -1,3 +1,4 @@
+h5py-stubs
 pandas-stubs
 pyarrow-stubs
 scipy-stubs; python_version >= '3.10'
